requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.39.15
botocore==1.39.15
certifi==2025.7.14
cffi==1.17.1
h11==0.16.0
idna==3.10
jmespath==1.0.1
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.3.1
pyarrow==21.0.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
s3transfer==0.13.1
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
